investment-agent/requirements.txt
# Core dependencies
yfinance>=0.2.14
pandas>=1.5.3
numpy>=1.24.0
requests>=2.28.2
python-dotenv>=1.0.0

# Data Analysis
statsmodels>=0.13.5
scikit-learn>=1.2.2

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.0

# API Integration
anthropic>=0.7.0  # For Claude API and MCP

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development Tools
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.3.0